% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_helpers.R
\name{check.fit}
\alias{check.fit}
\title{Check if a fit object is generated and if a weight has been set...}
\usage{
check.fit()
}
\description{
Check if a fit object is generated and if a weight has been set...
}
