% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_spear.R
\name{cv.evaluate}
\alias{cv.evaluate}
\title{Evaluate cv SPEAR object}
\usage{
cv.evaluate(
  nlambda = 100,
  calculate.factor.contributions = TRUE,
  max_iter = 10000,
  multinomial_loss = "deviance"
)
}
\arguments{
\item{nlambda}{Number of lambdas (defaults to 100)}

\item{calculate.factor.contributions}{Calculate factor contributions? When `$params$family == "multinomial"` or `"ordinal"` can save time to put `FALSE`. Defaults to `TRUE`.}

\item{max_iter}{Maximum number of iterations (defaults to 10000)}

\item{multinomial_loss}{Type of loss for when `$params$family == "multinomial"`. Can be `"deviance"` (default) or `"misclassification"`}
}
\description{
Evaluate cv SPEAR object
}
