% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cv.evaluation}
\alias{cv.evaluation}
\title{We consider two types the model:
(1) yhat = (U * b) * chat (refit = F)
(2) yhat = U * bhat (refit = T)}
\usage{
cv.evaluation(
  fitted.obj,
  X,
  Y,
  Z,
  family,
  nclasses,
  pattern_samples,
  pattern_features,
  nlambda = 100,
  factor_contribution = F,
  max_iter = 10000
)
}
\arguments{
\item{fitted.obj}{Fitted object from cv.spear}

\item{X}{Feature array.}

\item{Y}{Response array.}

\item{rule}{Rule to use for picking cross validation model: "min" or "1se".}

\item{standardize}{Whether to standardize the data.}

\item{alpha}{alpha = 1 corresponds to lasso and alpha = 0 corresponds to ridge.}
}
\description{
We consider two types the model:
(1) yhat = (U * b) * chat (refit = F)
(2) yhat = U * bhat (refit = T)
}
