% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spear.R
\name{cv.spear}
\alias{cv.spear}
\title{Cross-fit of SPEAR}
\usage{
cv.spear(
  X,
  Xobs,
  Y,
  Yobs,
  Z,
  family,
  nclasses,
  ws,
  num_factors,
  functional_path,
  foldid = foldid,
  pattern_samples = NULL,
  pattern_features = NULL,
  inits_type = "pca",
  warm_up = 100,
  max_iter = 1000,
  thres_elbo = 0.01,
  thres_count = 5,
  thres_factor = 1e-08,
  print_out = 10,
  a0 = 0.01,
  b0 = 0.01,
  a1 = sqrt(nrow(X)),
  b1 = sqrt(nrow(X)),
  a2 = sqrt(nrow(X)),
  b2 = sqrt(nrow(X)),
  robust_eps = 1/(sqrt(nrow(X))),
  inits_post_mu = NULL,
  seed = 1,
  crossYonly = F,
  numCores = NULL,
  run.debug = FALSE
)
}
\arguments{
\item{X}{Feature matrix.}

\item{Y}{Response matrix (can be multidimensional for gaussian data).}

\item{ws}{A vector of weights that you want to try out, default is ws = c(0).}

\item{functional_path}{Grouping structure.}

\item{foldid}{CV foldid}

\item{inits_type}{Initialization type, can be None, pca, sparsepac.}

\item{num.factors}{Number of factors estimated.}

\item{warm_up:}{warm up iterations for the inference.}

\item{max_iter:}{max number of iterations.}

\item{thres_elbo:}{if EBLO increase by less than thres_elbo, clock +1; otherwise, the clock is reset to 0.}

\item{thres_count:}{stop if clock has reached thres_count.}

\item{print_out:}{print out the progress for every print_out iterations.}

\item{a0:}{hyper parametr, no need to tune usually.}

\item{b0:}{hyper parametr, no need to tune usually.}

\item{a1:}{hyper parametr, no need to tune usually.}

\item{b1:}{hyper parametr, no need to tune usually.}

\item{a2:}{hyper parametr, no need to tune usually.}

\item{b2:}{hyper parametr, no need to tune usually.}

\item{seed:}{random seed number.}

\item{run.debug:}{debug?}
}
\description{
Cross-fit of SPEAR
}
