% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_helpers.R
\name{get.contributions}
\alias{get.contributions}
\title{Get factor contributions for a SPEARobject.}
\usage{
get.contributions(do.X = TRUE, do.Y = TRUE, do.Y.pvals = FALSE)
}
\arguments{
\item{do.X}{Calculate contributions for X (train)? Defaults to `TRUE`}

\item{do.Y}{Calculate contributions for Y? Defaults to `TRUE`}

\item{do.Y.pvals}{Include p.values for y? Defaults to `FALSE`}
}
\description{
Get factor contributions for a SPEARobject.
}
