% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_helpers.R
\name{get.factor.scores}
\alias{get.factor.scores}
\title{Get factor scores from a SPEARobject for a particular dataset `data`.}
\usage{
get.factor.scores(data = "train", cv = FALSE)
}
\arguments{
\item{data}{Which dataset to use? Can be any dataset listed under `$data$____`. Defaults to `"train"`.}

\item{cv}{If `data = "train"`, get factor scores generated from `$run.cv.spear`? If `$run.spear` was used or if `data != "train"` this parameter is ignored. Defaults to `FALSE`. NOTE: There is an element of randomness if the factor scores are not correlated with the response, so it is recomended to view the factor scores with `cv = FALSE`}
}
\description{
Get factor scores from a SPEARobject for a particular dataset `data`.
}
