% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_helpers.R
\name{get.features}
\alias{get.features}
\title{Get features from a SPEARobject.}
\usage{
get.features(
  rank = "probability",
  factors = NULL,
  datasets = NULL,
  coefficient.cutoff = 0,
  probability.cutoff = 0.95,
  correlation = "none"
)
}
\arguments{
\item{rank}{How to rank features? Defaults to `"probability"` (joint.probability) followed by projection.coefficient. Can also be "`coefficient`" (only projection.coefficient magnitude).}

\item{factors}{Which factors to return features for? Accepts integers (i.e. `factors = c(1,2,4)` or `factors = 1`) Defaults to `NULL` (all).}

\item{datasets}{Which datasets? Check names via `names(SPEARobject$data$train$Xlist)`. Defaults to `NULL` (all)}

\item{coefficient.cutoff}{What projection.coefficient value to use as the cutoff? Defaults to 0}

\item{probability.cutoff}{What joint.probability value to use as the cutoff? Defaults to .95}

\item{correlation}{Return columns with factor score correlations? Accepts `"spearman"`, `"pearson"`, `"kendall"`, or "`none`" (default)}
}
\description{
Get features from a SPEARobject.
}
