% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_helpers.R
\name{get.misclassification}
\alias{get.misclassification}
\title{Get misclassification error rates and confusion matrices from a SPEARobject and a dataset `data`.}
\usage{
get.misclassification(data = "train", cv = TRUE)
}
\arguments{
\item{data}{Which dataset to use? Can be any dataset listed under `$data$____`. Defaults to `"train"`.}

\item{cv}{If `data = "train"`, get factor scores generated from `$run.cv.spear`? If `$run.spear` was used or if `data != "train"` this parameter is ignored. Defaults to `TRUE`.}
}
\description{
Get misclassification error rates and confusion matrices from a SPEARobject and a dataset `data`.
}
