% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_helpers.R
\name{get.predictions}
\alias{get.predictions}
\title{Get predictions from a SPEARobject and a dataset `data`.}
\usage{
get.predictions(data = "train", cv = TRUE)
}
\arguments{
\item{data}{Which dataset to use? Can be any dataset listed under `$data$____`. Defaults to `"train"`.}

\item{cv}{If `data = "train"`, get factor scores generated from `$run.cv.spear`? If `$run.spear` was used or if `data != "train"` this parameter is ignored. Defaults to `TRUE`.}
}
\description{
Get predictions from a SPEARobject and a dataset `data`.
}
