% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit3.R
\name{get_SPEAR_model}
\alias{get_SPEAR_model}
\title{Get a SPEAR model for a specific weight 'w'. The value of 'w' can be specified (i.e. w = 1, w = 0) or the "overall"/"best" weight from}
\usage{
get_SPEAR_model(SPEARobj, w = "overall", w.method = "sd")
}
\arguments{
\item{SPEARobj}{SPEAR object (returned from \code{run_cv_spear})}

\item{w.method}{How to choose best weight? Options include \code{"min"} (lowest mean CV error, default) and \code{"sd"} (choose a higher weight within 1 standard deviation of the CV errors).}
}
\value{
A named vector of weights (or weight indices from \code{SPEARobj$params$weights} if \code{return.as.index} == \code{TRUE}).
}
\description{
Get a SPEAR model for a specific weight 'w'. The value of 'w' can be specified (i.e. w = 1, w = 0) or the "overall"/"best" weight from
}
\examples{
SPEAR.get_best_weights(SPEARobj, w.method = "min", return.overall = TRUE)
SPEAR.get_best_weights(SPEARobj, w.method = "sd")
}
