% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{impute.z}
\alias{impute.z}
\title{Impute the rest of a missing matrix (Z). Called internally.}
\usage{
impute.z(Z, method = NULL)
}
\arguments{
\item{method}{Which imputation method?}
}
\description{
Impute the rest of a missing matrix (Z). Called internally.
}
