% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_SPEARobject.R
\name{load.SPEARobject}
\alias{load.SPEARobject}
\title{Load a SPEARobject. Will return an R6 class SPEARobject used for the "SPEAR" package.}
\usage{
load.SPEARobject(file = NULL)
}
\arguments{
\item{file}{Where the SPEARobject Rds file is located. Defaults to NULL}
}
\description{
Load a SPEARobject. Will return an R6 class SPEARobject used for the "SPEAR" package.
}
