% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spear.R
\name{make_spear_object}
\alias{make_spear_object}
\title{Make a SPEARobject. Will return an R6 class SPEARobject used for the "SPEAR" package.}
\usage{
make_spear_object(
  X = NULL,
  Y = NULL,
  Z = NULL,
  family = "gaussian",
  weights.case = NULL,
  weights.x = NULL,
  weights.y = NULL,
  num_factors = 5,
  inits_type = "pca",
  inits_post_mu = NULL,
  sparsity_upper = 0.5,
  warm_up = 100,
  max_iter = 1000,
  thres_elbo = 0.01,
  thres_count = 5,
  thres_factor = 1e-08,
  print_out = 100,
  seed = 123,
  a0 = NULL,
  b0 = NULL,
  a1 = NULL,
  b1 = NULL,
  a2 = NULL,
  b2 = NULL,
  L0 = NULL,
  L1 = NULL,
  L2 = NULL,
  robust_eps = NULL,
  remove.formatting = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{X}{Assay matrix.}

\item{Y}{Response matrix (can be multidimensional for gaussian data).}

\item{Z}{Complete feature matrix (usually the features are the imputed version of X, other features are attached to the end).}
}
\description{
Make a SPEARobject. Will return an R6 class SPEARobject used for the "SPEAR" package.
}
