% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{preparation}
\alias{preparation}
\title{Create the data from lists, and create the functional pathway list.}
\usage{
preparation(
  Y,
  X,
  family,
  pattern_samples = NULL,
  pattern_assays = NULL,
  path.type = "assay",
  bx = NULL,
  other.path = NULL
)
}
\arguments{
\item{Y}{Response array.}

\item{X}{description}

\item{refit}{Whether to refit the model with elasticNet. If FALSE, the coefficients are from the spear model.}

\item{rule}{Rule to use for picking cross validation model: "min" or "1se".}

\item{standardize}{Whether to standardize the data.}

\item{alpha}{alpha = 1 corresponds to lasso and alpha = 0 corresponds to ridge.}
}
\description{
Create the data from lists, and create the functional pathway list.
}
