% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_helpers.R
\name{print.out}
\alias{print.out}
\title{Print out a variety of summary information about a SPEARobject}
\usage{
\method{print}{out}(type = NULL, remove.formatting = FALSE, quiet = FALSE)
}
\arguments{
\item{type}{Which type of summary to print? Can be "data". Defaults to NULL.}

\item{remove.formatting}{Remove text color/bold font face. Defaults to FALSE.}

\item{quiet}{Do not print anything. Defaults to FALSE.}
}
\description{
Print out a variety of summary information about a SPEARobject
}
\examples{
SPEARobj <- make_spear_model(...)

SPEARobj$run.spear()

}
