% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_spear.R
\name{run.cv.spear}
\alias{run.cv.spear}
\title{Run CV SPEAR}
\usage{
run.cv.spear(
  fold.ids = NULL,
  num.folds = NULL,
  only.cross.Y = FALSE,
  num.cores = NULL
)
}
\description{
Run CV SPEAR
}
\examples{
SPEARobj <- make.SPEARobject(...)

SPEARobj$run.cv.spear()

}
