% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spear.R
\name{run.spear}
\alias{run.spear}
\title{SPEAR - SuPervised Bayes fActor for Multi-omics
imports:}
\usage{
run.spear()
}
\description{
SPEAR - SuPervised Bayes fActor for Multi-omics
imports:
}
\examples{
SPEARobj <- make_spear_model(...)

SPEARobj$run.spear()

}
