% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{run_cv_spear}
\alias{run_cv_spear}
\title{Create the data from lists, and create the functional pathway list.}
\usage{
run_cv_spear(
  X,
  Y,
  Z = NULL,
  Xobs = NULL,
  Yobs = NULL,
  foldid = NULL,
  weights = NULL,
  family = "gaussian",
  inits.type = "pca",
  num.factors = NULL,
  seed = NULL,
  scale.x = TRUE,
  scale.y = TRUE,
  num.folds = 5,
  warmup.iterations = NULL,
  max.iterations = NULL,
  elbo.threshold = NULL,
  elbo.threshold.count = NULL,
  cv.nlambda = 100,
  print.out = 100,
  save.model = TRUE,
  save.path = NULL,
  save.name = NULL,
  run.debug = FALSE,
  robust_eps = NULL,
  sparsity_upper = 0.1,
  L0 = 1,
  factor_contribution = TRUE
)
}
\arguments{
\item{X}{description}

\item{Y}{description}

\item{Z}{description}

\item{Xobs}{description}

\item{Yobs}{description}

\item{foldid}{description}

\item{weights}{description}

\item{family}{description}

\item{inits.type}{description}

\item{num.factors}{description}

\item{seed}{description}

\item{scale.x}{description}

\item{scale.y}{description}

\item{num.folds}{description}

\item{warmup.iterations}{description}

\item{max.iterations}{description}

\item{elbo.threshold}{description}

\item{elbo.threshold.count}{description}

\item{cv.nlambda}{description}

\item{print.out}{description}

\item{save.model}{description}

\item{save.path}{description}

\item{save.name}{description}

\item{run.debug}{debug?}

\item{sparsity_upper}{sparsity_upper, defaults to .5}

\item{L0}{parameter}

\item{factor_contribution}{Calculate factor contributions? Defaults to TRUE}
}
\description{
Create the data from lists, and create the functional pathway list.
}
