% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_helpers.R
\name{save.model}
\alias{save.model}
\title{Save a SPEARobject}
\usage{
save.model(file = NULL)
}
\arguments{
\item{file}{Name and path for the stored SPEARobject. Defaults to NULL, which is (workingdirectory)/SPEARobject__date__}
}
\description{
Save a SPEARobject
}
\examples{
SPEARobj <- make.SPEARobject(...)

SPEARobj$run.cv.spear()$cv.evaluation()

SPEARobj$save.model(path = getwd(), file = "SPEARmodel.rds")
}
