% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit3.R
\name{SPEAR.color_text}
\alias{SPEAR.color_text}
\title{Color text for output in the terminal}
\usage{
SPEAR.color_text(text, fg = "black", bg = NULL)
}
\arguments{
\item{text}{A string to be colored}

\item{fg}{Foreground color}

\item{bg}{Background color}
}
\value{
Text that has been formatted.
}
\description{
Color text for output in the terminal
}
\examples{
SPEAR.color_text("Error", fg = "red", bg = NULL)
}
