% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit3.R
\name{SPEAR.get_best_features}
\alias{SPEAR.get_best_features}
\title{Get the best features from the variance cutoff parameter}
\usage{
SPEAR.get_best_features(
  SPEARmodel,
  factor = 1,
  var.cutoff = 0.5,
  forecast = "out.of.sample"
)
}
\arguments{
\item{SPEARmodel}{A SPEAR model (returned from get_SPEAR_model)}

\item{coefficient.cutoff}{Cutoff for coefficient magnitude. If .1 is provided, abs(coefficients) >= .1 are returned. Defaults to .01}

\item{probability.cutoff}{Cutoff for posterior selection probability. Ranges from 0 - 1. Defaults to .5}
}
\description{
Get the best features from the variance cutoff parameter
}
