% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit.R
\name{SPEAR.get_color_scheme}
\alias{SPEAR.get_color_scheme}
\title{Get a color scheme for a SPEAR plot}
\usage{
SPEAR.get_color_scheme(SPEARobj)
}
\arguments{
\item{SPEARobj}{SPEAR object (returned from run_cv_spear)}
}
\description{
Get a color scheme for a SPEAR plot
}
