% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit.R
\name{SPEAR.get_cv_loss}
\alias{SPEAR.get_cv_loss}
\title{Get loss for each value of \code{w} from training a SPEAR model.}
\usage{
SPEAR.get_cv_loss(SPEARobj, include.sd = FALSE)
}
\arguments{
\item{SPEARobj}{SPEAR object (returned from \code{run_cv_spear})}

\item{include.sd}{Return standard deviation of loss as well? Defaults to \code{FALSE}.}
}
\value{
A matrix with the first column being each \code{w} trained on and a column for every response. If \code{include.sd == TRUE}, a list of matrices, named \code{mean} and \code{sd}.
}
\description{
Get loss for each value of \code{w} from training a SPEAR model.
}
\examples{
SPEAR.get_cv_loss(SPEARobj)
SPEAR.get_cv_loss(SPEARobj, include.sd = TRUE)
}
