% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit.R
\name{SPEAR.get_cv_predictions}
\alias{SPEAR.get_cv_predictions}
\title{Get CV predictions}
\usage{
SPEAR.get_cv_predictions(SPEARobj, w = "best")
}
\arguments{
\item{SPEARobj}{SPEAR object (returned from run_cv_spear)}
}
\description{
Get CV predictions
}
