% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit3.R
\name{SPEAR.get_data_summary}
\alias{SPEAR.get_data_summary}
\title{Check that data are normally distributed and scaled properly for SPEAR}
\usage{
SPEAR.get_data_summary(Xlist = NULL, Y = NULL, plot.xlist.pvalues = TRUE)
}
\arguments{
\item{Xlist}{A list of matrices of omics data. Defaults to NULL.}

\item{Y}{A data.frame of response variables. Defaults to NULL.}

\item{plot.xlist.pvalues}{Should the function return plots of p.values from the shapiro-wilk normality tests? Defaults to TRUE}
}
\value{
NULL (will print out results)
}
\description{
Check that data are normally distributed and scaled properly for SPEAR
}
\examples{
SPEAR.get_data_summary(Xlist, Y)
SPEAR.get_data_summary(SPEARmodel$data$xlist, SPEARmodel$data$Y)
SPEAR.get_data_summary(SPEARobject$data$xlist, SPEARobject$data$Y)
}
