% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit.R
\name{SPEAR.get_factor_contributions}
\alias{SPEAR.get_factor_contributions}
\title{Get factor contributions to Y}
\usage{
SPEAR.get_factor_contributions(SPEARobj, threshold = 0.01)
}
\arguments{
\item{SPEARobj}{SPEAR object (returned from run_cv_spear)}
}
\description{
Get factor contributions to Y
}
