% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit.R
\name{SPEAR.get_factor_features}
\alias{SPEAR.get_factor_features}
\title{Get factor features}
\usage{
SPEAR.get_factor_features(SPEARobj, cutoff = 0.5, threshold = 0.01)
}
\arguments{
\item{SPEARobj}{SPEAR object (returned from run_cv_spear)}
}
\description{
Get factor features
}
