% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit.R
\name{SPEAR.get_factor_scores}
\alias{SPEAR.get_factor_scores}
\title{Get factor scores from a trained SPEAR object on training/test data}
\usage{
SPEAR.get_factor_scores(
  SPEARobj,
  w = "best",
  X = NULL,
  w.method = "min",
  scale.x = FALSE
)
}
\arguments{
\item{SPEARobj}{SPEAR object (returned from \code{run_cv_spear})}

\item{w}{Weight for SPEAR model. Defaults to \code{"best"}, choosing the best weight per response. Can also be \code{"overall"} (choosing the weight with the best overall mean cross-validated error), or one of the weights used to train SPEAR (\code{SPEARobj$params$weights})}

\item{X}{List of omics matrices to be used for prediction. Defaults to \code{NULL} (use training data stored within SPEAR)}

\item{w.method}{How to choose best weight? Options include \code{"min"} (lowest mean CV error, default) and \code{"sd"} (choose a higher weight within 1 standard deviation of the CV errors).}

\item{scale.x}{Should \code{X} be scaled (only used if new \code{X} is supplied). Defaults to \code{FALSE}}
}
\value{
A list of predictions from the SPEAR model. One list for each response predicted, as well as \code{predictions}, \code{probabilites}, and \code{w} used.
}
\description{
Get factor scores from a trained SPEAR object on training/test data
}
\examples{
SPEAR.get_factor_scores(SPEARobj)
SPEAR.get_factor_scores(SPEARobj, w = 1)
}
