% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit3.R
\name{SPEAR.get_feature_table}
\alias{SPEAR.get_feature_table}
\title{Get a data.frame of all features that pass cutoffs from the chosen model:}
\usage{
SPEAR.get_feature_table(
  SPEARmodel,
  rank = "probability",
  factors = NULL,
  omics = NULL,
  coefficient.cutoff = 0,
  probability.cutoff = 0.95,
  correlation = "none",
  var.cutoff = 0.7,
  num.features = NULL,
  forecast = "out.of.sample",
  correlation.method = "pearson"
)
}
\arguments{
\item{SPEARmodel}{SPEAR model (returned from \code{get_SPEAR_model})}

\item{rank}{How should features be ranked? "probability" (default) ranks first by joint.probability, followed by projection.coefficient magnitude. Features will be selected by coefficient.cutoff and probability.cutoff (see below) "variance" returns a set number of features that explain a single factor. Features are selected one at a time, choosing the feature with the highest correlation, calculating variance explained, regressing that feature out, and re-ordering features based on their correlation. Will continue until an amount of total variance explained reaches var.exp.cutoff or a number of features is reached (num.features).}

\item{factors}{A vector of integers representing factors. If NULL, will return features for all factors. Defaults to NULL (all factors)}

\item{omics}{A vector of names of datasets (see names(SPEARmodel$data$xlist)). Defaults to NULL (all omics)}

\item{coefficient.cutoff}{Cutoff for coefficient magnitude. If .1 is provided, abs(coefficients) >= .1 are returned. Defaults to .01}

\item{probability.cutoff}{Cutoff for posterior selection probability. Ranges from 0 - 1. Defaults to .5}

\item{correlation}{Which method for correlation? Defaults to "none", but can be "pearson", "spearman", or "kendall"}
}
\value{
A data.frame with features that passed cutoffs set by the parameters.
}
\description{
Get a data.frame of all features that pass cutoffs from the chosen model:
}
\examples{
SPEAR.get_feature_table(SPEARmodel)

# Get specific features for one omic:
omic.name <- names(SPEARmodel$data$X)[1]
SPEAR.get_feature_table(SPEARmodel, factors = 1, omics = omic.name)

# Get features that pass cutoffs:
SPEAR.get_feature_table(SPEARmodel, coefficient.cutoff = .2, probability.cutoff = .5)
}
