% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit3.R
\name{SPEAR.get_predictions}
\alias{SPEAR.get_predictions}
\title{Use a SPEARmodel to predict response for new samples (Xlist)}
\usage{
SPEAR.get_predictions(SPEARmodel, Xlist = NULL)
}
\arguments{
\item{SPEARmodel}{SPEAR model (returned from \code{get_SPEAR_model})}
}
\value{
A matrix of predictions (out.of.sample)
}
\description{
Use a SPEARmodel to predict response for new samples (Xlist)
}
\examples{
SPEAR.get_predictions(SPEARmodel)
SPEAR.get_predictions(SPEARmodel, Xlist = Xlist)
}
