% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit3.R
\name{SPEAR.get_top_features}
\alias{SPEAR.get_top_features}
\title{Get the top features from the variance cutoff parameter. First, takes the top feature (ranked by joint probability, then projection coefficients)}
\usage{
SPEAR.get_top_features(
  SPEARmodel,
  factor = 1,
  var.cutoff = 0.7,
  num.features = NULL,
  coefficient.cutoff = 0,
  probability.cutoff = 0.95,
  forecast = "out.of.sample",
  correlation.method = "pearson"
)
}
\arguments{
\item{SPEARmodel}{A SPEAR model (returned from get_SPEAR_model)}

\item{factor}{Which factor? Must be a single integer (i.e. 1, 2, 4). Defaults to 1}

\item{var.cutoff}{How much variance needs to be explained? Must be between 0 - 1. Defaults to .7}

\item{num.features}{Rather than looping until 'var.cutoff' is met, give a predetermined number of features (i.e. 50, 100, ...). If 'num.features' is provided, will ignore 'var.cutoff'. Defaults to NULL (will use 'var.cutoff' instead)}

\item{coefficient.cutoff}{Cutoff for coefficient magnitude. If .1 is provided, abs(coefficients) >= .1 are returned. Defaults to .01}

\item{probability.cutoff}{Cutoff for posterior selection probability. Ranges from 0 - 1. Defaults to .5}

\item{forecast}{Which factor scores to use? A string with "out.of.sample" or "in.sample". Defaults to "out.of.sample".}

\item{correlation.method}{Which method to be used to find the next best feature? Can be "spearman", "pearson", and "kendall". Defaults to "pearson" (fastest)}
}
\description{
Get the top features from the variance cutoff parameter. First, takes the top feature (ranked by joint probability, then projection coefficients)
}
