% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit3.R
\name{SPEAR.plot_biplot}
\alias{SPEAR.plot_biplot}
\title{Plot a biplot of the chosen Factor scores and features}
\usage{
SPEAR.plot_biplot(
  SPEARmodel,
  x.factor = 1,
  y.factor = 2,
  groups = NULL,
  forecast = "out.of.sample"
)
}
\arguments{
\item{SPEARmodel}{A SPEAR model (returned from get_SPEAR_model)}

\item{groups}{A named vector of a grouping variable, where names(groups) = rownames(SPEARmodel$data$X). Names are required to match to ensure correct representation of the data.}

\item{forecast}{Which probabilities to use? A string with "out.of.sample" or "in.sample". Defaults to "out.of.sample".}

\item{x}{Which Factor to show on the x-axis? String of a Factor name (i.e. "Factor1", default).}

\item{y}{Which Factor to show on the y-axis? String of a Factor name (i.e. "Factor2", default).}
}
\description{
Plot a biplot of the chosen Factor scores and features
}
