% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit3.R
\name{SPEAR.plot_class_probabilities}
\alias{SPEAR.plot_class_probabilities}
\title{Plot class probabilities for a SPEARmodel (of family "binomial", "categorical", or "ordinal")}
\usage{
SPEAR.plot_class_probabilities(
  SPEARmodel,
  groups = NULL,
  forecast = "out.of.sample"
)
}
\arguments{
\item{SPEARmodel}{SPEAR model (returned from \code{get_SPEAR_model})}

\item{groups}{A named vector of a grouping variable, where names(groups) = rownames(SPEARmodel$data$X). Names are required to match to ensure correct representation of the data.}

\item{forecast}{Which probabilities to use? A string with "out.of.sample" or "in.sample". Defaults to "out.of.sample".}
}
\value{
A plot showing in-sample vs. out-of-sample predictions per response
}
\description{
Plot class probabilities for a SPEARmodel (of family "binomial", "categorical", or "ordinal")
}
\examples{
SPEAR.plot_class_probabilities(SPEARmodel)

groups <- SPEARmodel$params$foldid
names(groups) <- rownames(SPEARmodel$data$Y)
SPEAR.plot_class_probabilities(SPEARmodel, groups = groups, forecast = "in.sample")
}
