% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit3.R
\name{SPEAR.plot_correlation}
\alias{SPEAR.plot_correlation}
\title{Plot a correlation matrix for the selected features}
\usage{
SPEAR.plot_correlation(
  SPEARmodel,
  x = "Factor1",
  y = "Factor2",
  groups = NULL,
  method = "spearman",
  forecast = "out.of.sample"
)
}
\arguments{
\item{SPEARmodel}{A SPEAR model (returned from get_SPEAR_model)}

\item{x}{What to show on the x-axis? Can be a string of a Factor name (i.e. "Factor1", default) or string of a feature name (see colnames(SPEARmodel$data$X)).}

\item{y}{What to show on the y-axis? Can be a string of a Factor name (i.e. "Factor2", default) or string of a feature name (see colnames(SPEARmodel$data$X)).}

\item{groups}{A named vector of a grouping variable, where names(groups) = rownames(SPEARmodel$data$X). Names are required to match to ensure correct representation of the data.}

\item{method}{Method for correlation. Can be 'pearson', 'kendall', or 'spearman'. Defaults to 'spearman'.}

\item{forecast}{Which probabilities to use? A string with "out.of.sample" or "in.sample". Defaults to "out.of.sample".}
}
\description{
Plot a correlation matrix for the selected features
}
