% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit.R
\name{SPEAR.plot_cv_prediction_errors}
\alias{SPEAR.plot_cv_prediction_errors}
\title{Plot factor scores per subject}
\usage{
SPEAR.plot_cv_prediction_errors(
  SPEARobj,
  show.w.labels = TRUE,
  show.min.w.line = TRUE
)
}
\arguments{
\item{SPEARobj}{SPEAR object (returned from run_cv_spear)}

\item{show.w.labels}{Label points with their weights? Defaults to TRUE}

\item{show.min.w.line}{Draw a dashed line at the lowest mean cv error? Defaults to TRUE}
}
\description{
Plot factor scores per subject
}
