% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit.R
\name{SPEAR.plot_cv_predictions}
\alias{SPEAR.plot_cv_predictions}
\title{Plot CV predictions}
\usage{
SPEAR.plot_cv_predictions(SPEARobj, ncol = NULL, nrow = 1)
}
\arguments{
\item{SPEARobj}{SPEAR object (returned from run_cv_spear)}
}
\description{
Plot CV predictions
}
