% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit3.R
\name{SPEAR.plot_factor_contributions}
\alias{SPEAR.plot_factor_contributions}
\title{Plot factor contributions for X and Y}
\usage{
SPEAR.plot_factor_contributions(
  SPEARmodel,
  threshold = 0.01,
  show.labels = TRUE,
  show.irrelevant = FALSE
)
}
\arguments{
\item{SPEARmodel}{SPEAR model (returned from get_SPEAR_model)}

\item{threshold}{Threshold value of contribution for a factor to be relevant. Defaults to .01.}

\item{show.labels}{Show the contributions as geom_text on the plot? Defaults to TRUE}

\item{show.irrelevant}{Show all contributions, even those below the threshold? Defaults to FALSE}
}
\description{
Plot factor contributions for X and Y
}
