% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit3.R
\name{SPEAR.plot_factor_grid}
\alias{SPEAR.plot_factor_grid}
\title{Plot factor scores for a SPEARmodel}
\usage{
SPEAR.plot_factor_grid(
  SPEARmodel,
  groups = NULL,
  forecast = "out.of.sample",
  Xlist = NULL,
  fit.line = TRUE
)
}
\arguments{
\item{SPEARmodel}{SPEAR model (returned from \code{get_SPEAR_model})}

\item{groups}{A named vector of a grouping variable, where names(groups) = rownames(SPEARmodel$data$X). Names are required to match to ensure correct representation of the data.}

\item{forecast}{Which probabilities to use? A string with "out.of.sample" or "in.sample". Defaults to "out.of.sample".}

\item{Xlist}{A list of matrices of testing data (must match the same dimensions as the data used to train the SPEARmodel (SPEARmodel$data$xlist)). Defaults to NULL (use training data).}

\item{fit.line}{Add linear line of best fit? Defaults to TRUE}
}
\value{
A plot showing in-sample vs. out-of-sample predictions per response
}
\description{
Plot factor scores for a SPEARmodel
}
\examples{
SPEAR.plot_factor_grid(SPEARmodel, jitter.points = FALSE)

groups <- SPEARmodel$params$foldid
names(groups) <- rownames(SPEARmodel$data$Y)
SPEAR.plot_factor_grid(SPEARmodel, groups = groups, forecast = "in.sample")
}
