% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit.R
\name{SPEAR.plot_factor_loadings}
\alias{SPEAR.plot_factor_loadings}
\title{Plot factor loadings}
\usage{
SPEAR.plot_factor_loadings(
  SPEARobj,
  w = "best",
  w.method = "sd",
  threshold = 0.01,
  response.name = NULL,
  plot.per.omic = TRUE,
  scale.per.omic = TRUE,
  show.feature.names = FALSE,
  plot.irrelevant.factors = FALSE
)
}
\arguments{
\item{SPEARobj}{SPEAR object (returned from run_cv_spear)}

\item{w}{Weight for SPEAR. Defaults to "best", choosing the best weight per response. Can also be "overall" (choosing the weight with the best overall mean cross-validated error), or one of the weights used to train SPEAR (SPEARobj$params$weights)}

\item{threshold}{Threshold value of contribution for a factor to be relevant. Defaults to .01.}

\item{response.name}{Which response variable? Check colnames(SPEARobj$data$Y) for options. Defaults to 1st column (NULL)}

\item{plot.per.omic}{Return an individual plot per omic (each dataset in SPEARobj$data$xlist)? Defaults to TRUE}

\item{scale.per.omic}{Return an individual plot per omic (each dataset in SPEARobj$data$xlist)? Defaults to TRUE}

\item{show.feature.names}{Show row names (features) for factor loadings? Defaults to FALSE}

\item{plot.irrelevant.factors}{Plot loadings for factors with a contribution lower than the threshold? Defaults to FALSE}
}
\description{
Plot factor loadings
}
