% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit3.R
\name{SPEAR.plot_factor_probabilities}
\alias{SPEAR.plot_factor_probabilities}
\title{Plot posterior selection probabilities for a SPEARmodel}
\usage{
SPEAR.plot_factor_probabilities(
  SPEARmodel,
  type = "joint",
  plot.per.omic = TRUE,
  plot.log = TRUE
)
}
\arguments{
\item{SPEARmodel}{SPEAR model (returned from \code{get_SPEAR_model})}

\item{type}{Which probabilities to plot? Options include "marginal", "regression", and "joint" (default)}
}
\value{
A plot with the factor probabilities of the SPEAR model
}
\description{
Plot posterior selection probabilities for a SPEARmodel
}
\examples{
SPEAR.plot_factor_probabilities(SPEARmodel)
}
