% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit.R
\name{SPEAR.plot_factor_scores}
\alias{SPEAR.plot_factor_scores}
\title{Plot factor scores}
\usage{
SPEAR.plot_factor_scores(
  SPEARobj,
  w = "overall",
  groups = NULL,
  factors = NULL,
  return.as.list = FALSE,
  include.legend = TRUE
)
}
\arguments{
\item{SPEARobj}{SPEAR object (returned from run_cv_spear)}
}
\description{
Plot factor scores
}
