% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit3.R
\name{SPEAR.plot_feature_correlation}
\alias{SPEAR.plot_feature_correlation}
\title{Plot a correlation matrix for the selected features}
\usage{
SPEAR.plot_feature_correlation(
  SPEARmodel,
  feature.table,
  method = "spearman",
  show.feature.names = TRUE,
  show.correlation.values = FALSE,
  cluster.features = TRUE
)
}
\arguments{
\item{SPEARmodel}{A SPEAR model (returned from get_SPEAR_model)}

\item{feature.table}{A table of features from SPEAR.get_feature_table}

\item{method}{Method for correlation. Can be 'pearson', 'kendall', or 'spearman'. Defaults to 'spearman'.}

\item{show.feature.names}{Should the feature names be shown on the axes? Defaults to TRUE}

\item{show.correlation.values}{Should the (rounded) correlation values be shown? Defaults to FALSE}

\item{cluster.features}{Cluster the features in the heatmap? Defaults to TRUE}
}
\description{
Plot a correlation matrix for the selected features
}
