% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit.R
\name{SPEAR.plot_feature_summary}
\alias{SPEAR.plot_feature_summary}
\title{Plot factor features}
\usage{
SPEAR.plot_feature_summary(
  SPEARobj,
  w = "best",
  factor = NULL,
  response.name = NULL,
  threshold = 0.01,
  probability.cutoff = 0.5,
  coefficient.cutoff = 0,
  show.top.cutoff = 50,
  group.by.omic = FALSE
)
}
\arguments{
\item{SPEARobj}{SPEAR object (returned from run_cv_spear)}

\item{w}{Weight for SPEAR. Defaults to "best", choosing the best weight per response. Can also be "overall" (choosing the weight with the best overall mean cross-validated error), or one of the weights used to train SPEAR (SPEARobj$params$weights)}

\item{factor}{Which factors to plot? Defaults to NULL (all relevant factors).}

\item{response.name}{Which response variable? Check colnames(SPEARobj$data$Y) for options. Defaults to 1st column (NULL)}

\item{threshold}{Threshold value of contribution for a factor to be relevant. Defaults to .01.}

\item{probability.cutoff}{Posterior selection probability cutoff (0 - 1) for a feature to be selected. Defaults to .5}

\item{coefficient.cutoff}{Coefficient cutoff (positive) for a feature to be selected. Will use +/-. Defaults to .01}

\item{show.top.cutoff}{How many features to show? Set to NULL to show all. Defaults to 50.}

\item{group.by.omic}{Group features from the same omic together? Defaults to FALSE}
}
\description{
Plot factor features
}
