% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit3.R
\name{SPEAR.plot_feature_table}
\alias{SPEAR.plot_feature_table}
\title{Plot the top features from the variance cutoff parameter. First, takes the top feature (ranked by joint probability, then projection coefficients)}
\usage{
SPEAR.plot_feature_table(
  SPEARmodel,
  feature.table,
  plot.type = "summary",
  correlation.method = "pearson",
  show.feature.names = TRUE,
  show.correlation.values = FALSE,
  cluster.features = TRUE
)
}
\arguments{
\item{SPEARmodel}{A SPEAR model (returned from get_SPEAR_model)}

\item{feature.table}{A feature table returned by SPEAR.get_feature_table.}

\item{plot.type}{Which plot to show? Options include "summary" (default), plots joint probabilities vs. projection coefficients, "variance" - plot variance explained per feature in a Factor.}
}
\description{
Plot the top features from the variance cutoff parameter. First, takes the top feature (ranked by joint probability, then projection coefficients)
}
