% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit.R
\name{SPEAR.plot_ordinal_class_probabilities}
\alias{SPEAR.plot_ordinal_class_probabilities}
\title{Plot ordinal class probabilities}
\usage{
SPEAR.plot_ordinal_class_probabilities(
  SPEARobj,
  X = NULL,
  Y = NULL,
  w = "best",
  response.name = NULL,
  scale.x = FALSE
)
}
\arguments{
\item{SPEARobj}{SPEAR object (returned from run_cv_spear)}

\item{X}{List of omics matrices to be used for prediction. Defaults to NULL (use training data stored within SPEAR)}

\item{Y}{Responses for subjects (rows) in X. Defaults to NULL (use training data stored within SPEAR)}

\item{w}{Weight for SPEAR. Defaults to "best", choosing the best weight per response. Can also be "overall" (choosing the weight with the best overall mean cross-validated error), or one of the weights used to train SPEAR (SPEARobj$params$weights)}

\item{response.name}{Which response variable? Check colnames(SPEARobj$data$Y) for options. Defaults to 1st column (NULL)}

\item{scale.x}{Should X be scaled (only used if X is supplied). Defaults to FALSE}
}
\description{
Plot ordinal class probabilities
}
