% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit3.R
\name{SPEAR.plot_overfit}
\alias{SPEAR.plot_overfit}
\title{Use a SPEARmodel to predict response for new samples (Xlist)}
\usage{
SPEAR.plot_overfit(SPEARmodel, groups = NULL)
}
\arguments{
\item{SPEARmodel}{SPEAR model (returned from \code{get_SPEAR_model})}

\item{groups}{A named vector of a grouping variable, where names(groups) = rownames(SPEARmodel$data$X). Names are required to match to ensure correct representation of the data.}
}
\value{
A plot showing in-sample vs. out-of-sample predictions per response
}
\description{
Use a SPEARmodel to predict response for new samples (Xlist)
}
\examples{
SPEAR.get_cv_predictions(SPEARmodel)
}
