% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit3.R
\name{SPEAR.plot_variance_per_feature}
\alias{SPEAR.plot_variance_per_feature}
\title{Plot variance per feature}
\usage{
SPEAR.plot_variance_per_feature(
  SPEARmodel,
  factor = 1,
  num.features = 50,
  coefficient.cutoff = 0,
  probability.cutoff = 0,
  forecast = "out.of.sample"
)
}
\arguments{
\item{SPEARmodel}{A SPEAR model (returned from get_SPEAR_model)}

\item{coefficient.cutoff}{Cutoff for coefficient magnitude. If .1 is provided, abs(coefficients) >= .1 are returned. Defaults to .01}

\item{probability.cutoff}{Cutoff for posterior selection probability. Ranges from 0 - 1. Defaults to .5}
}
\description{
Plot variance per feature
}
