% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit.R
\name{SPEAR.predict_new_samples}
\alias{SPEAR.predict_new_samples}
\title{Predict response values for new samples.}
\usage{
SPEAR.predict_new_samples(SPEARobj, X, scale.x = TRUE)
}
\arguments{
\item{SPEARobj}{SPEAR object (returned from run_cv_spear)}
}
\description{
Predict response values for new samples.
}
