% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit.R
\name{SPEAR.predict_ordinal_classes}
\alias{SPEAR.predict_ordinal_classes}
\title{Predict ordinal classes}
\usage{
SPEAR.predict_ordinal_classes(
  SPEARobj,
  X = NULL,
  w = "overall",
  return.probabilities = TRUE,
  scale.x = FALSE
)
}
\arguments{
\item{SPEARobj}{SPEAR object (returned from run_cv_spear)}

\item{X}{List of omics matrices to be used for prediction. Defaults to NULL (use training data stored within SPEAR)}

\item{w}{Weight for SPEAR. Defaults to "best", choosing the best weight per response. Can also be "overall" (choosing the weight with the best overall mean cross-validated error), or one of the weights used to train SPEAR (SPEARobj$params$weights)}

\item{return.probabilities}{Return ordinal class probabilities? Defaults to TRUE}

\item{scale.x}{Should X be scaled (only used if X is supplied). Defaults to FALSE}
}
\description{
Predict ordinal classes
}
