% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_spear.R
\name{spear}
\alias{spear}
\title{SPEAR - SuPervised Bayes fActor for Multi-omics
imports:}
\usage{
spear(
  X = NULL,
  Y = NULL,
  Z = NULL,
  Xobs = NULL,
  Yobs = NULL,
  weights_case = NULL,
  silence = NULL
)
}
\description{
SPEAR - SuPervised Bayes fActor for Multi-omics
imports:
}
