% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARinteractive.R
\name{SPEAR.run_interactive}
\alias{SPEAR.run_interactive}
\title{Get best SPEAR weights per response Y}
\usage{
SPEAR.run_interactive(SPEARobj)
}
\arguments{
\item{SPEARobj}{SPEAR object (returned from \code{run_cv_spear})}
}
\description{
Get best SPEAR weights per response Y
}
\examples{
SPEAR.run_interactive(SPEARobject)
}
