% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit3.R
\name{SPEAR.set_color_scheme}
\alias{SPEAR.set_color_scheme}
\title{Set a color scheme for a SPEAR plot}
\usage{
SPEAR.set_color_scheme(SPEARobject, X.vector = NULL, Y.vector = NULL)
}
\arguments{
\item{SPEARobject}{SPEAR object (returned from get_SPEAR_model)}
}
\description{
Set a color scheme for a SPEAR plot
}
