% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R, R/spear.R
\name{update.dimnames}
\alias{update.dimnames}
\title{Update the dimension names for all SPEARobject matrices. Used internally.}
\usage{
\method{update}{dimnames}(call = "fit")

\method{update}{dimnames}(call = "fit")
}
\description{
Update the dimension names for all SPEARobject matrices. Used internally.

Update the dimension names for all SPEARobject matrices. Used internally.
}
