% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_helpers.R
\name{color.text}
\alias{color.text}
\title{Color text for output in the terminal}
\usage{
color.text(text, fg = "black", bg = NULL)
}
\arguments{
\item{text}{A string to be colored}

\item{fg}{Foreground color}

\item{bg}{Background color}
}
\value{
Text that has been formatted.
}
\description{
Color text for output in the terminal
}
\examples{
color.text("Error", fg = "red", bg = NULL)
}
