% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit.R
\name{SPEAR.get_best_weights}
\alias{SPEAR.get_best_weights}
\title{Get best SPEAR weights per response Y}
\usage{
SPEAR.get_best_weights(SPEARobj, include.best.overall = FALSE)
}
\arguments{
\item{SPEARobj}{SPEAR object (returned from run_cv_spear)}
}
\description{
Get best SPEAR weights per response Y
}
