% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit3.R
\name{SPEAR.get_best_weights}
\alias{SPEAR.get_best_weights}
\title{Get best SPEAR weights per response Y}
\usage{
SPEAR.get_best_weights(
  SPEARobj,
  w.method = "sd",
  return.overall = FALSE,
  return.as.index = FALSE
)
}
\arguments{
\item{SPEARobj}{SPEAR object (returned from \code{run_cv_spear})}

\item{w.method}{How to choose best weight? Options include \code{"min"} (lowest mean CV error, default) and \code{"sd"} (choose a higher weight within 1 standard deviation of the CV errors).}

\item{return.overall}{Should the best overall weight be returned? Defaults to \code{FALSE} (where each response can have a different best weight).}

\item{return.as.index}{Return the index (from SPEARobj$params$weights) rather than the weight itself? Defaults to \code{FALSE}.}
}
\value{
A named vector of weights (or weight indices from \code{SPEARobj$params$weights} if \code{return.as.index} == \code{TRUE}).
}
\description{
Get best SPEAR weights per response Y
}
\examples{
SPEAR.get_best_weights(SPEARobj, w.method = "min", return.overall = TRUE)
SPEAR.get_best_weights(SPEARobj, w.method = "sd")
}
