% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit.R
\name{SPEAR.get_cv_prediction_error}
\alias{SPEAR.get_cv_prediction_error}
\title{Get CV prediction errors}
\usage{
SPEAR.get_cv_prediction_error(SPEARobj, show.all.w = FALSE, verbose = FALSE)
}
\arguments{
\item{SPEARobj}{SPEAR object (returned from run_cv_spear)}
}
\description{
Get CV prediction errors
}
