% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit.R
\name{SPEAR.get_cv_predictions}
\alias{SPEAR.get_cv_predictions}
\title{Get CV predictions}
\usage{
SPEAR.get_cv_predictions(SPEARobj, w = "best")
}
\arguments{
\item{SPEARobj}{SPEAR object (returned from run_cv_spear)}

\item{w}{Weight for SPEAR. Defaults to "best", choosing the best weight per response. Can also be "overall" (choosing the weight with the best overall mean cross-validated error), or one of the weights used to train SPEAR (SPEARobj$params$weights)}
}
\description{
Get CV predictions
}
