% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit.R
\name{SPEAR.get_factor_contributions}
\alias{SPEAR.get_factor_contributions}
\title{Get factor contributions to both X and Y}
\usage{
SPEAR.get_factor_contributions(SPEARobj, w = "best", threshold = 0.01)
}
\arguments{
\item{SPEARobj}{SPEAR object (returned from run_cv_spear)}

\item{w}{Weight for SPEAR. Defaults to "best", choosing the best weight per response. Can also be "overall" (choosing the weight with the best overall mean cross-validated error), or one of the weights used to train SPEAR (SPEARobj$params$weights)}

\item{threshold}{Threshold value of contribution for a factor to be relevant. Defaults to .01.}
}
\description{
Get factor contributions to both X and Y
}
