% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit.R
\name{SPEAR.get_factor_scores}
\alias{SPEAR.get_factor_scores}
\title{Get factor scores}
\usage{
SPEAR.get_factor_scores(SPEARobj, w = "best")
}
\arguments{
\item{SPEARobj}{SPEAR object (returned from run_cv_spear)}
}
\description{
Get factor scores
}
