% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit3.R
\name{SPEAR.get_factor_scores}
\alias{SPEAR.get_factor_scores}
\title{Use a SPEARmodel to get factor scores for new samples (Xlist)}
\usage{
SPEAR.get_factor_scores(SPEARmodel, Xlist = NULL)
}
\arguments{
\item{SPEARmodel}{SPEAR model (returned from \code{get_SPEAR_model})}

\item{Xlist}{A list of matrices of testing data (must match the same dimensions as the data used to train the SPEARmodel (SPEARmodel$data$xlist)). Defaults to NULL (returning out-of-sample cv predictions for training data).}
}
\value{
A matrix of factor scores (out.of.sample)
}
\description{
Use a SPEARmodel to get factor scores for new samples (Xlist)
}
\examples{
SPEAR.get_factor_scores(SPEARmodel)
SPEAR.get_factor_scores(SPEARmodel, Xlist = Xlist)
}
