% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit3.R
\name{SPEAR.get_feature_table}
\alias{SPEAR.get_feature_table}
\title{Get a data.frame of all features that pass cutoffs from the chosen model:}
\usage{
SPEAR.get_feature_table(
  SPEARmodel,
  factors = NULL,
  omics = NULL,
  coefficient.cutoff = 0,
  probability.cutoff = 0.95,
  method = "spearman"
)
}
\arguments{
\item{SPEARmodel}{SPEAR model (returned from \code{get_SPEAR_model})}

\item{factors}{A vector of integers representing factors. If NULL, will return features for all factors. Defaults to NULL (all factors)}

\item{omics}{A vector of names of datasets (see names(SPEARmodel$data$xlist)). Defaults to NULL (all omics)}

\item{coefficient.cutoff}{Cutoff for coefficient magnitude. If .1 is provided, abs(coefficients) >= .1 are returned. Defaults to .01}

\item{probability.cutoff}{Cutoff for posterior selection probability. Ranges from 0 - 1. Defaults to .5}

\item{method}{Which method for correlation? Defaults to "spearman"}
}
\value{
A data.frame with features that passed cutoffs set by the parameters.
}
\description{
Get a data.frame of all features that pass cutoffs from the chosen model:
}
\examples{
SPEAR.get_feature_table(SPEARmodel)

# Get specific features for one omic:
omic.name <- names(SPEARmodel$data$X)[1]
SPEAR.get_feature_table(SPEARmodel, factors = 1, omics = omic.name)

# Get features that pass cutoffs:
SPEAR.get_feature_table(SPEARmodel, coefficient.cutoff = .2, probability.cutoff = .5)
}
