% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit3.R
\name{SPEAR.plot_class_predictions}
\alias{SPEAR.plot_class_predictions}
\title{Plot ordinal class predictions}
\usage{
SPEAR.plot_class_predictions(
  SPEARmodel,
  forecast = "out.of.sample",
  show.true.distribution = TRUE
)
}
\arguments{
\item{SPEARmodel}{A SPEAR model (returned from get_SPEAR_model)}

\item{forecast}{Which probabilities to use? A string with "out.of.sample" or "in.sample". Defaults to "out.of.sample".}

\item{show.true.distribution}{Plot the true distribution as well? Defaults to TRUE}
}
\description{
Plot ordinal class predictions
}
