% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit3.R
\name{SPEAR.plot_factor_coefficients}
\alias{SPEAR.plot_factor_coefficients}
\title{Plot factor coefficients for a SPEARmodel}
\usage{
SPEAR.plot_factor_coefficients(SPEARmodel, type = "regression")
}
\arguments{
\item{SPEARmodel}{SPEAR model (returned from \code{get_SPEAR_model})}

\item{type}{Which coefficients to plot? Options include "marginal" and "regression" (default)}
}
\value{
A plot with the factor loadings of the SPEAR model
}
\description{
Plot factor coefficients for a SPEARmodel
}
\examples{
SPEAR.plot_factor_coefficients(SPEARmodel)
}
