% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit.R
\name{SPEAR.plot_factor_coefficients}
\alias{SPEAR.plot_factor_coefficients}
\title{Plot factor features}
\usage{
SPEAR.plot_factor_coefficients(
  SPEARobj,
  w = "best",
  factor = NULL,
  response.name = NULL,
  threshold = 0.01,
  probability.cutoff = 0.5,
  coefficient.cutoff = 0.01
)
}
\arguments{
\item{SPEARobj}{SPEAR object (returned from run_cv_spear)}

\item{w}{Weight for SPEAR. Defaults to "best", choosing the best weight per response. Can also be "overall" (choosing the weight with the best overall mean cross-validated error), or one of the weights used to train SPEAR (SPEARobj$params$weights)}

\item{threshold}{Threshold value of contribution for a factor to be relevant. Defaults to .01.}

\item{probability.cutoff}{Posterior selection probability cutoff (0 - 1) for a feature to be selected. Defaults to .5}

\item{coefficient.cutoff}{Coefficient cutoff (positive) for a feature to be selected. Will use +/-. Defaults to .01}
}
\description{
Plot factor features
}
