% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit.R
\name{SPEAR.plot_factor_contributions}
\alias{SPEAR.plot_factor_contributions}
\title{Plot factor contributions to Y}
\usage{
SPEAR.plot_factor_contributions(
  SPEARobj,
  threshold = 0.01,
  show.labels = TRUE,
  show.irrelevant = FALSE
)
}
\arguments{
\item{SPEARobj}{SPEAR object (returned from run_cv_spear)}
}
\description{
Plot factor contributions to Y
}
