% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit.R
\name{SPEAR.plot_factor_loadings}
\alias{SPEAR.plot_factor_loadings}
\title{Plot factor loadings}
\usage{
SPEAR.plot_factor_loadings(
  SPEARobj,
  threshold = 0.01,
  plot.per.omic = FALSE,
  return.list = FALSE,
  show.feature.names = FALSE,
  plot.irrelevant.factors = FALSE
)
}
\arguments{
\item{SPEARobj}{SPEAR object (returned from run_cv_spear)}
}
\description{
Plot factor loadings
}
