% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit3.R
\name{SPEAR.plot_factor_loadings}
\alias{SPEAR.plot_factor_loadings}
\title{Plot factor scores for a SPEARmodel}
\usage{
SPEAR.plot_factor_loadings(SPEARmodel, plot.per.omic = TRUE)
}
\arguments{
\item{SPEARmodel}{SPEAR model (returned from \code{get_SPEAR_model})}

\item{plot.per.omic}{Should a bar with colors for each omic be plotted along side the loadings? Defaults to TRUE}
}
\value{
A plot with the factor loadings of the SPEAR model
}
\description{
Plot factor scores for a SPEARmodel
}
\examples{
SPEAR.plot_factor_loadings(SPEARmodel)
}
