% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit3.R
\name{SPEAR.plot_factor_scores}
\alias{SPEAR.plot_factor_scores}
\title{Plot factor scores for a SPEARmodel}
\usage{
SPEAR.plot_factor_scores(
  SPEARmodel,
  groups = NULL,
  forecast = "out.of.sample",
  jitter.points = TRUE
)
}
\arguments{
\item{SPEARmodel}{SPEAR model (returned from \code{get_SPEAR_model})}

\item{groups}{A named vector of a grouping variable, where names(groups) = rownames(SPEARmodel$data$X). Names are required to match to ensure correct representation of the data.}

\item{forecast}{Which probabilities to use? A string with "out.of.sample" or "in.sample". Defaults to "out.of.sample".}

\item{jitter.points}{Use geom_jitter vs. geom_point? Defaults to TRUE}
}
\value{
A plot showing in-sample vs. out-of-sample predictions per response
}
\description{
Plot factor scores for a SPEARmodel
}
\examples{
SPEAR.plot_factor_scores(SPEARmodel, jitter.points = FALSE)

groups <- SPEARmodel$params$foldid
names(groups) <- rownames(SPEARmodel$data$Y)
SPEAR.plot_factor_scores(SPEARmodel, groups = groups, forecast = "in.sample")
}
