% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEARkit3.R
\name{SPEAR.plot_feature_summary}
\alias{SPEAR.plot_feature_summary}
\title{Plot posterior seleciton probabilities and coefficients for features that pass the provided cutoffs.}
\usage{
SPEAR.plot_feature_summary(
  SPEARmodel,
  factors = NULL,
  omics = NULL,
  coefficient.cutoff = 0.01,
  probability.cutoff = 0.5,
  sort.by = "probability",
  max.per.factor = 10
)
}
\arguments{
\item{SPEARmodel}{SPEAR model (returned from \code{get_SPEAR_model})}

\item{factors}{A vector of integers representing factors. If NULL, will return features for all factors. Defaults to NULL (all factors)}

\item{omics}{A vector of names of datasets (see names(SPEARmodel$data$xlist)). Defaults to NULL (all omics)}

\item{coefficient.cutoff}{Cutoff for coefficient magnitude. If .1 is provided, abs(coefficients) >= .1 are returned. Defaults to .01}

\item{probability.cutoff}{Cutoff for posterior selection probability. Ranges from 0 - 1. Defaults to .5}

\item{sort.by}{How to order the features? "probability" ranks by posterior selection probability first, then coefficient. "coefficient" ranks purely by coefficient. Defaults to "probability"}

\item{max.per.factor}{Maximum number of features per factor? Defaults to 10 (may be more due to overlapping features)}
}
\value{
A plot of features that pass the provided cutoffs.
}
\description{
Plot posterior seleciton probabilities and coefficients for features that pass the provided cutoffs.
}
\examples{
SPEAR.plot_feature_summary(SPEARmodel)

# Plot specific features for one omic:
omic.name <- names(SPEARmodel$data$X)[1]
SPEAR.plot_feature_summary(SPEARmodel, factors = 1, omics = omic.name)

# Plot features that pass cutoffs:
SPEAR.plot_feature_summary(SPEARmodel, coefficient.cutoff = .2, probability.cutoff = .5)
}
